Split out input/event related methods into partial class
